refactor: extract findmnt parsing into _findmnt.py submodule
