+++ projects/storage-device-managers/src/storage_device_managers/__init__.py
@@ -4,15 +4,15 @@ import secrets
 import string
 import tempfile
 import typing as t
-from collections import defaultdict
 from collections.abc import Iterator
 from importlib import metadata
 from pathlib import Path
 from types import SimpleNamespace
 from uuid import UUID, uuid4
 
-import msgspec
 import shell_interface as sh
+
+from storage_device_managers._findmnt import MountOptions, get_mounted_devices
 
 try:
     from loguru import logger  # type: ignore[import, unused-ignore]
@@ -50,7 +50,6 @@ __all__ = [
     "unmount_device",
 ]
 
-MountOptions = frozenset[str]
 ValidFileSystems = t.Literal["btrfs", "ext4"]
 
 
@@ -403,53 +402,6 @@ def is_mounted(device: Path) -> bool:
         logger.info(f"Kein Mountpunkt für Speichermedium {device} gefunden.")
         return False
     return True
-
-
-class _FindmntFilesystem(msgspec.Struct):
-    source: str
-    target: str
-    options: str
-    fstype: str = ""
-
-
-class _FindmntOutput(msgspec.Struct):
-    filesystems: list[_FindmntFilesystem]
-
-
-def get_mounted_devices() -> t.Mapping[str, t.Mapping[Path, MountOptions]]:
-    """Get all mounted devices
-
-    This function will parse the output of `findmnt -l --json` and return everything
-    that is mounted to somewhere. The returned mapping maps device names (i.e. mount
-    sources) to their destinations and mount options.
-
-    Since a source can be mounted to multiple (e.g. /dev/sda1 can be mounted to
-    /home/{user1,user2}/Videos), the value of the mapping is another mapping. This inner
-    mapping maps mount destinations to their mount options.
-
-    Returns:
-    --------
-    t.Mapping[str, t.Mapping[Path, MountOptions]]
-        A mapping that maps mount sources (i.e. device names) to their
-        destinations and mount options.
-
-    Example Return Value:
-    ---------------------
-    {
-        "/dev/nvme0n1p2": {
-            Path("/boot"): frozenset({"rw", "relatime"}),
-            Path("/media/backup"): frozenset({"rw", "relatime", "compress=zstd:3"}),
-        },
-    }
-    """
-    raw = sh.run_cmd(cmd=["findmnt", "-l", "--json"], capture_output=True)
-    parsed = msgspec.json.decode(raw.stdout, type=_FindmntOutput)
-    mount_points: dict[str, dict[Path, MountOptions]] = defaultdict(dict)
-    for fs in parsed.filesystems:
-        dest = Path(fs.target)
-        options: MountOptions = frozenset(fs.options.split(","))
-        mount_points[fs.source][dest] = options
-    return dict(mount_points)
 
 
 def sync_device(device: Path) -> None:
